+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - It contained only one public method `simbak.backup.backup()`, and wasn’t able to scale well.
     - Moved the one public method to `simbak/__init__.py` so that the method signature is smaller `simbak.backup()`
     - This can be the one and only quick backup method to use, which will perform a standard backup. It should meet the needs for most people, and keeps it simple. Other more complex backup solutions will be in the new `simbak/backup` package.
+- Added unit tests.
 
 # v0.1.2
 - Critical error with logging fixed, logging should now work normally.